feature(canvas): move gradient mode controls to top toolbar; remove popover mode buttons and group clip+mode cluster
