+++ invokeai/frontend/web/src/features/controlLayers/components/Tool/ToolGradientModeToggle.tsx
@@ -1,0 +1,45 @@
+import { ButtonGroup, IconButton, Tooltip } from '@invoke-ai/ui-library';
+import { useAppDispatch, useAppSelector } from 'app/store/storeHooks';
+import { selectGradientType, settingsGradientTypeChanged } from 'features/controlLayers/store/canvasSettingsSlice';
+import { memo, useCallback } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { GradientLinearIcon, GradientRadialIcon } from './GradientIcons';
+
+export const ToolGradientModeToggle = memo(() => {
+  const { t } = useTranslation();
+  const gradientType = useAppSelector(selectGradientType);
+  const dispatch = useAppDispatch();
+
+  const onLinearClick = useCallback(() => dispatch(settingsGradientTypeChanged('linear')), [dispatch]);
+  const onRadialClick = useCallback(() => dispatch(settingsGradientTypeChanged('radial')), [dispatch]);
+
+  return (
+    <ButtonGroup isAttached size="sm">
+      <Tooltip label={t('controlLayers.gradient.linear', { defaultValue: 'Linear' })}>
+        <IconButton
+          aria-label={t('controlLayers.gradient.linear', { defaultValue: 'Linear' })}
+          icon={<GradientLinearIcon />}
+          colorScheme={gradientType === 'linear' ? 'invokeBlue' : 'base'}
+          variant="solid"
+          w="30px"
+          h="30px"
+          onClick={onLinearClick}
+        />
+      </Tooltip>
+      <Tooltip label={t('controlLayers.gradient.radial', { defaultValue: 'Radial' })}>
+        <IconButton
+          aria-label={t('controlLayers.gradient.radial', { defaultValue: 'Radial' })}
+          icon={<GradientRadialIcon />}
+          colorScheme={gradientType === 'radial' ? 'invokeBlue' : 'base'}
+          variant="solid"
+          w="30px"
+          h="30px"
+          onClick={onRadialClick}
+        />
+      </Tooltip>
+    </ButtonGroup>
+  );
+});
+
+ToolGradientModeToggle.displayName = 'ToolGradientModeToggle';
